refactor: decouple operator commands from TUI
